# openapi/notes-api-v1.yaml
...
openapi: 3.0.0
info:
  title: OSM Notes API
  version: '0.1.0'
  description: |
    REST API for OSM Notes Analytics and Ingestion. Provides programmatic access to
    user profiles, country analytics, notes, and real-time metrics.

    Responses that contain data from OpenStreetMap or derivatives include an
    `attribution` field (text and url) so clients can display proper attribution
    (e.g. "© OpenStreetMap contributors").
  contact:
    name: OSM Notes Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://notes-api.osm.lat
    description: Production server
security:
  - UserAgent: []
components:
  securitySchemes:
    UserAgent:
      type: apiKey
      in: header
      name: User-Agent
      description: 'User-Agent header required for all requests. Format: AppName/Version (Contact)'
  schemas:
    Error:
      type: object
      required: [error, message, statusCode]
      properties:
        error:
          type: string
          description: Error type name
          example: Bad Request
        message:
          type: string
          description: Human-readable error message
          example: Invalid note ID
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
...
    GlobalAnalytics:
      type: object
      properties:
        dimension_global_id:
          type: integer
          example: 1
        history_whole_open:
          type: integer
          example: 1000000
        history_whole_closed:
          type: integer
          example: 800000
        currently_open_count:
          type: integer
          nullable: true
          example: 200000
        resolution_rate:
          type: number
          format: float
          nullable: true
          example: 80.0
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 250
        total_pages:
          type: integer
          example: 13
...
    TrendEntry:
      type: object
      properties:
        year:
          type: string
        open:
          type: integer
        closed:
          type: integer
    TrendsResult:
      type: object
      properties:
        type:
          type: string
          enum: [users, countries, global]
        entity_id:
          type: integer
        entity_name:
          type: string
          nullable: true
        trends:
          type: array
          items:
            $ref: '#/components/schemas/TrendEntry'
        working_hours:
          type: array
          items:
            type: integer
          description: Array of 168 numbers (24h * 7 days)
paths:
...
  /notes-api/v1/analytics/global:
    get:
      summary: Get global analytics and statistics
      tags: [Analytics]
      security:
        - UserAgent: []
      responses:
        '200':
          description: Global analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GlobalAnalytics'
        '404':
          description: Global analytics not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes-api/v1/analytics/comparison:
    get:
      summary: Compare users or countries by their metrics
      tags: [Analytics]
      security:
        - UserAgent: []
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [users, countries]
          description: Type of entities to compare
        - in: query
          name: ids
          required: true
          schema:
            type: string
          description: Comma-separated list of IDs (max 10)
      responses:
        '200':
          description: Comparison result with metrics per entity
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes-api/v1/analytics/trends:
    get:
      summary: Get temporal trends for users, countries, or global
      tags: [Analytics]
      security:
        - UserAgent: []
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [users, countries, global]
          description: Type of entity for trends
        - in: query
          name: user_id
          schema:
            type: integer
          description: User ID (required if type is users)
        - in: query
          name: country_id
          schema:
            type: integer
          description: Country ID (required if type is countries)
      responses:
        '200':
          description: Trends data with temporal analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsResult'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /notes-api/v1/hashtags:
    get:
      summary: Get list of hashtags with usage counts
      tags: [Hashtags]
      security:
        - UserAgent: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Results per page
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (by count)
      responses:
        '200':
          description: List of hashtags with counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        hashtag:
                          type: string
                        count:
                          type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes-api/v1/hashtags/{hashtag}:
    get:
      summary: Get details for a specific hashtag
      tags: [Hashtags]
      security:
        - UserAgent: []
      parameters:
        - in: path
          name: hashtag
          required: true
          schema:
            type: string
          description: Hashtag name (without #)
      responses:
        '200':
          description: Hashtag details with users and countries using it
          content:
            application/json:
              schema:
                type: object
                properties:
                  hashtag:
                    type: string
                  users_count:
                    type: integer
                  countries_count:
                    type: integer
        '400':
          description: Invalid hashtag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Hashtag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'